+++ TrabalhoFinal/include/MRoleta.h
@@ -8,7 +8,7 @@ class MRoleta : public Maquina
 {
 
     public:
-        MRoleta(int _id, string _nome, int _x, int _y, int _premio, float _prob, string _tipo, int _aposta);
+        MRoleta(int _id, string _nome, int _x, int _y, float _premio, float _prob, string _tipo, int _aposta);
         virtual ~MRoleta();
 
     protected:
