+++ TrabalhoFinal/src/MSlot.cpp
@@ -2,7 +2,7 @@
 #include <stdlib.h>     /* srand, rand */
 #include <time.h>       /* time */
 
-MSlot::MSlot(int _id, string _nome, int _x, int _y, int _premio, float _prob, string _tipo, int _aposta)
+MSlot::MSlot(int _id, string _nome, int _x, int _y, float _premio, float _prob, string _tipo, int _aposta)
     : Maquina(_id, _nome, _x, _y, _premio, _prob, _tipo, _aposta)
 {
     //ctor
